data-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
openai==1.12.0
pydantic==2.6.0
python-dotenv==1.0.1
playwright==1.40.0
playwright-stealth==1.0.6
